pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "appgen"
version = "0.1.1"
description = "A modern project generator CLI for web frameworks"
authors = [
    { name = "Ratish jain", email = "ratishjain6@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
keywords = ["cli", "generator", "web", "framework", "project"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0"
]
requires-python = ">=3.7"

[project.scripts]
appgen = "appgen.cli:app"

[tool.setuptools]
packages = ["appgen", "generator"]
py-modules = ["config"]

[tool.setuptools.package-data]
"generator" = ["../templates/**/*"]
"appgen" = ["../appgen.config.yaml"]

[project.urls]
Homepage = "https://github.com/yourusername/appgen"
Repository = "https://github.com/yourusername/appgen"
Issues = "https://github.com/yourusername/appgen/issues"